Implement workspace/workspaceDispatch, cleanup, move away from redux (#407)
* Implement workspace/workspaceDispatch
Fixes #406

* add ApiConnection.ts

* fix python compile

* Workspace view in jupyter (#409)

* WIP: Implement WorkspaceViewJp

* fix WorkspaceViewJp

* fix jupyter compile

* Add example2.ipynb

* fix memory issue on install test and misc updates

* fix github action

* trigger github action
