+++ jupyterlab/labbox_ephys_widgets_jp/labbox_ephys_widgets_jp/__init__.py
@@ -7,6 +7,7 @@
 from ._version import __version__, version_info
 from .create_recording_view import create_recording_view
 from .create_sorting_view import create_sorting_view
+from .create_workspace_view import create_workspace_view, WorkspaceView
 
 from .request_handlers import load_jupyter_server_extension
