+++ jupyterlab/labbox_ephys_widgets_jp/src/index.ts
@@ -1,6 +1,7 @@
 // Copyright (c) Jeremy Magland
 // Distributed under the terms of the Modified BSD License.
 
-export * from './pluginWidgets';
+export * from './SortingViewJp';
 export * from './version';
+export * from './WorkspaceViewJp';
