+++ src/extensions/old/__init__.py
@@ -1,2 +1,1 @@
 from .averagewaveformsold import *
-from .curation import *
