--- src/extensions/old/curation/__init__.py
@@ -1,1 +1,0 @@
-from .curation import *
