+++ src/registerExtensions.ts
@@ -12,7 +12,6 @@ import { activate as activateelectrodegeometry } from './extensions/electrodegeo
 import { activate as activatepythonsnippets } from './extensions/pythonsnippets/pythonsnippets'
 import { activate as activatesnippets } from './extensions/snippets/snippets'
 import { activate as activateexample } from './extensions/example/example'
-import { activate as activatecuration } from './extensions/old/curation/curation'
 import { activate as activateaveragewaveformsold } from './extensions/old/averagewaveformsold/averagewaveformsold'
 import { activate as activateclusters } from './extensions/clusters/clusters'
 import { activate as activateunitstable } from './extensions/unitstable/unitstable'
@@ -41,7 +40,6 @@ const registerExtensions = (context: ExtensionContext) => {
     activatepythonsnippets(context)
     activatesnippets(context)
     activateexample(context)
-    activatecuration(context)
     activateaveragewaveformsold(context)
     activateclusters(context)
     activateunitstable(context)
